+++ gq/gq.go
@@ -14,8 +14,6 @@ import (
 
 var GQ256 = jwa.SignatureAlgorithm("GQ256")
 
-const GQ256SecurityParameter = 256
-
 func init() {
 	jwa.RegisterSignatureAlgorithm(GQ256)
 }
@@ -60,7 +58,7 @@ type signerVerifier struct {
 
 // Creates a new SignerVerifier specifically for GQ256, meaning the security parameter is 256.
 func New256SignerVerifier(publicKey *rsa.PublicKey) (SignerVerifier, error) {
-	return NewSignerVerifier(publicKey, GQ256SecurityParameter)
+	return NewSignerVerifier(publicKey, 256)
 }
 
 // NewSignerVerifier creates a SignerVerifier from the RSA public key of the trusted third-party which creates
