+++ verifier/verifier.go
@@ -6,8 +6,11 @@ import (
 	"fmt"
 
 	"github.com/lestrrat-go/jwx/v2/jws"
+	"github.com/openpubkey/openpubkey/gq"
 	"github.com/openpubkey/openpubkey/pktoken"
 )
+
+var ErrNonGQUnsupported = fmt.Errorf("non-GQ signatures are not supported")
 
 type ProviderVerifier interface {
 	// Returns the OpenID provider issuer as seen in ID token e.g. "https://accounts.google.com"
@@ -43,6 +46,22 @@ func AddProviderVerifiers(verifiers ...ProviderVerifier) VerifierOpts {
 				return fmt.Errorf("provider verifier found with duplicate issuer: %s", verifier.Issuer())
 			}
 			v.providers[verifier.Issuer()] = verifier
+		}
+		return nil
+	}
+}
+
+type Check func(*Verifier, *pktoken.PKToken) error
+
+func GQOnly() Check {
+	return func(_ *Verifier, pkt *pktoken.PKToken) error {
+		alg, ok := pkt.ProviderAlgorithm()
+		if !ok {
+			return fmt.Errorf("missing provider algorithm header")
+		}
+
+		if alg != gq.GQ256 {
+			return ErrNonGQUnsupported
 		}
 		return nil
 	}
